added search bar, cleaned code

+++ contracts/Escrow.sol
@@ -13,7 +13,6 @@ interface IERC721 {
 }
 
 contract Escrow {
-    
     //state variables for actors
     address public lender;
     address public inspector;
@@ -23,37 +22,43 @@ contract Escrow {
     address public nftAddress;
 
     //mapping to store lsiting status of the nft
-    mapping(uint => bool) public isListed;
+    mapping(uint256 => bool) public isListed;
 
     //mapping for purchase price,escrow amount,buyer
-    mapping(uint => uint) public purchasePrice;
-    mapping(uint => uint) public escrowAmount;
-    mapping(uint => address) public buyer;
+    mapping(uint256 => uint256) public purchasePrice;
+    mapping(uint256 => uint256) public escrowAmount;
+    mapping(uint256 => address) public buyer;
 
     //mapping for inspection status
-    mapping(uint => bool) public inspectionPassed;
+    mapping(uint256 => bool) public inspectionPassed;
 
     //mapping for each actor approving sale
-    mapping(uint => mapping(address => bool)) public saleApproved;
+    mapping(uint256 => mapping(address => bool)) public saleApproved;
 
     //only seller can do this
-    modifier onlySeller(){
+    modifier onlySeller() {
         require(msg.sender == seller, "only seller can call this function");
         _;
     }
 
     //only buyer can do this
-    modifier onlyBuyer(uint _nftID){
-        require(msg.sender == buyer[_nftID], "only buyer can call this function");
+    modifier onlyBuyer(uint256 _nftID) {
+        require(
+            msg.sender == buyer[_nftID],
+            "only buyer can call this function"
+        );
         _;
     }
 
     //make the contract payable
-    receive() external payable{}
+    receive() external payable {}
 
     //only inpsector can do this
-    modifier onlyInspector(){
-        require(msg.sender == inspector,"only inspector can call this function");
+    modifier onlyInspector() {
+        require(
+            msg.sender == inspector,
+            "only inspector can call this function"
+        );
         _;
     }
 
@@ -63,24 +68,20 @@ contract Escrow {
         address payable _seller,
         address _inspector,
         address _lender
-        
-        ){
-        
+    ) {
         nftAddress = _nftAddress;
         seller = _seller;
         inspector = _inspector;
         lender = _lender;
-        
     }
 
     //move the nft from seller wallet to Escrow and list it
-    function list
-        (uint _nftID, 
-        uint _purchasePrice, 
-        uint _escrowAmount, 
+    function list(
+        uint256 _nftID,
+        uint256 _purchasePrice,
+        uint256 _escrowAmount,
         address _buyer
-        )public payable onlySeller{
-
+    ) public payable onlySeller {
         //we use the interface of erc721 to use the transferFrom function
         IERC721(nftAddress).transferFrom(msg.sender, address(this), _nftID);
 
@@ -94,29 +95,30 @@ contract Escrow {
     }
 
     //function for buyer to deposit earnest amount to contract
-    function depositEarnest(uint _nftID)public payable onlyBuyer(_nftID){
-        
+    function depositEarnest(uint256 _nftID) public payable onlyBuyer(_nftID) {
         require(msg.value >= escrowAmount[_nftID]);
     }
 
     //function get contract balacnce
-    function getBalance()public view returns(uint){
+    function getBalance() public view returns (uint256) {
         return address(this).balance;
     }
 
     //update inspection status, by only ispector
-    function updateInspectionPassed(uint _nftID, bool _isPassed)public onlyInspector{
+    function updateInspectionPassed(uint256 _nftID, bool _isPassed)
+        public
+        onlyInspector
+    {
         inspectionPassed[_nftID] = _isPassed;
     }
 
     //approve sale
-    function approveSale(uint _nftID)public{
+    function approveSale(uint256 _nftID) public {
         saleApproved[_nftID][msg.sender] = true;
     }
 
     //function to finalize sale
-    function finalizeSale(uint _nftID) public{
-        
+    function finalizeSale(uint256 _nftID) public {
         //inspection should be passed
         require(inspectionPassed[_nftID]);
 
@@ -132,7 +134,9 @@ contract Escrow {
         isListed[_nftID] = false;
 
         //transfer funds to seller
-        (bool success, ) = payable(seller).call{value : address(this).balance}("");
+        (bool success, ) = payable(seller).call{value: address(this).balance}(
+            ""
+        );
         require(success);
 
         //transfer nft to buyer
@@ -140,13 +144,11 @@ contract Escrow {
     }
 
     //function to cancel the sale
-    function cancelSale(uint _nftID)public{
-
+    function cancelSale(uint256 _nftID) public {
         //if insection failed transfer funds to buyer else send to seller
-        if(inspectionPassed[_nftID] == false){
+        if (inspectionPassed[_nftID] == false) {
             payable(buyer[_nftID]).transfer(address(this).balance);
-        }
-        else{
+        } else {
             payable(seller).transfer(address(this).balance);
         }
     }
