+++ contracts/RealEstate.sol
@@ -5,8 +5,7 @@ import "@openzeppelin/contracts/utils/Counters.sol";
 import "@openzeppelin/contracts/token/ERC721/ERC721.sol";
 import "@openzeppelin/contracts/token/ERC721/extensions/ERC721URIStorage.sol";
 
-contract RealEstate is ERC721URIStorage{
-
+contract RealEstate is ERC721URIStorage {
     using Counters for Counters.Counter;
     Counters.Counter private _tokenIds;
 
@@ -15,13 +14,12 @@ contract RealEstate is ERC721URIStorage{
 
     //create mint()
     //we are going to let someone mint with a tokenURI ie metadata
-    function mint(string memory tokenURI)public returns (uint){
-        
+    function mint(string memory tokenURI) public returns (uint256) {
         //default of tokenIds is 0, so we make it 1 and increment for every NFT created
         _tokenIds.increment();
 
         //get current value of tokenIds
-        uint newItemId = _tokenIds.current();
+        uint256 newItemId = _tokenIds.current();
 
         //mint() from erc721 contract
         _mint(msg.sender, newItemId);
@@ -30,12 +28,10 @@ contract RealEstate is ERC721URIStorage{
         _setTokenURI(newItemId, tokenURI);
 
         return newItemId;
-
     }
 
     //get total supply ,ie no of minted nfts
-    function totalSupply()public view returns(uint){
+    function totalSupply() public view returns (uint256) {
         return _tokenIds.current();
     }
-
 }
